refactor: split locationParser into ParserErrors and error message (#11653)
